src/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
attrs==21.4.0
autoflake==1.4
black==21.12b0
bpyutils==0.4.0
certifi==2021.10.8
chardet==4.0.0
charset-normalizer==2.0.12
click==8.1.1
colorclass==2.2.2
coreapi==2.3.3
coreschema==0.0.4
coverage==6.3.2
dj-rest-auth==2.2.4
Django==4.0.4
django-address==0.2.6
django-autocomplete-light==3.9.4
django-cors-headers==3.11.0
django-crispy-forms==1.14.0
django-extensions==3.1.5
django-filter==21.1
django-recurrence==1.11.1
djangorestframework==3.13.1
docopt==0.6.2
docutils==0.17.1
drf-spectacular==0.22.1
EasyProcess==1.1
freezegun==1.2.1
future==0.18.2
graphviz==0.20
gunicorn==20.1.0
idna==3.3
inflection==0.5.1
install==1.3.5
isort==5.10.1
itypes==1.2.0
Jinja2==3.1.1
jsonschema==4.4.0
MarkupSafe==2.1.1
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
pdfkit==1.0.0
pip-upgrader==1.4.15
platformdirs==2.5.1
probableparsing==0.0.1
psycopg2==2.9.3
pydot==1.4.2
pydot-ng==2.0.0
pydotplus==2.0.2
pyflakes==2.4.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-crfsuite==0.9.8
python-dateutil==2.8.2
pytz==2022.1
PyVirtualDisplay==3.0
PyYAML==5.3.1
regex==2022.3.15
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
terminaltables==3.1.10
toml==0.10.2
tomli==1.2.3
typed-ast==1.5.2
typing_extensions==4.1.1
uritemplate==4.1.1
urllib3==1.26.9
usaddress==0.5.10
python-dotenv==0.20.0
